Move Knob files
