+++ components/pages/example/KnobRow/KnobRow.tsx
@@ -1,13 +1,13 @@
 import { kebabCase } from 'lodash';
 import { Subtitle } from '@leafygreen-ui/typography';
-import { Knob } from './Knob';
 import { spacing } from '@leafygreen-ui/tokens';
 import { palette } from '@leafygreen-ui/palette';
 import { css } from '@leafygreen-ui/emotion';
 import { HTMLElementProps } from '@leafygreen-ui/lib';
-
 import InlineDefinition from '@leafygreen-ui/inline-definition';
-import { KnobType } from './utils';
+
+import { Knob } from '../Knob/Knob';
+import { KnobType } from '../utils';
 import { PropTooltipContent } from 'components/PropTooltipContent';
 
 const knobRowWrapperStyle = (darkMode: boolean) => css`
